requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-community==0.0.2
langchain-openai==0.0.2
langchain-text-splitters==0.0.1
langchain-core==0.1.7
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
tiktoken==0.5.1
faiss-cpu==1.7.4 